Reorganized Code
Removed the useEffect in the products page causing api call on each navigation to that page.
